bug: 53770 Disable mail if admin delegated login and admin mail is disabled, regardless of the mail enabled setting. Also hide all searches that only target mail items so that a search cannot end up showing mail items.
This required a lot of re-arranging of the ZmZimbraMail
startup sequence. But in doing so have eliminated the need
to pre-fetch some settings for use at the beginning of the
sequence before all the settings are initialized later.
Instead, all of the settings are initialized at the start
of the sequence from the get info response.

http://bugzilla.zimbra.com/show_bug.cgi?id=53770