Integrate change 275892 from zimbra to zimbra
-------------
Bug: 53898, 55793, 55990 - Optionally display addresses as friendly name
Optionally display the short form (display name) of an email address in the message header and in compose. Clicking on a bubble in compose will still select the full address, which is hidden, so a Copy (ctrl-C) operation will copy that to the clipboard. A tooltip shows the full address.
- AjxEmailAddress.toString() now takes an arg to have it return the short form (display name) rather than the full address.
- Fix a bug where dragging a bubble to an input with one bubble would always put it first.
- Don't draw rubber-band box if user clicks in non-empty INPUT, or if there are no bubbles to select.

http://bugzilla.zimbra.com/show_bug.cgi?id=53898
http://bugzilla.zimbra.com/show_bug.cgi?id=55793
http://bugzilla.zimbra.com/show_bug.cgi?id=55990