Integrate change 276525 from zimbra to zimbra
-------------
Bug 56377
The bug occurred by the null call of this._childDiv. By making sure
the invoked attribute is valid, the bug can be fixed.
The null of this._childDiv is because GetAdminExtensionZimletsResponse
do not return any valid zimlet attributes. The empty zimlet attributes
causes the div is not created in corresponding tree node.

http://bugzilla.zimbra.com/show_bug.cgi?id=56377