Integrate change 276950 from zimbra to zimbra
-------------
Bug:56860 - Right-click on a Contact :  "New Email" selection does not put email into the "To" field
The contacts code that gets an email string from the selected contacts relied on AjxEmailAddress::toString() not taking any args, but it now takes one for producing the short form. So we just separate out the handling of contact groups.
Also, if ZmAddressInputField::setValue() is handed a string that's not an email address, create a bubble for it but don't also put it in the INPUT.

http://bugzilla.zimbra.com/show_bug.cgi?id=56860