Integrate change 277572 from zimbra to zimbra
-------------
Bug 56731
changed the hidden select element to be a textarea for IE since the span there was somehow unselectable. No idea why, as the unselectable attribute does not seem to be applied to it (and in the header of email view the bubble text is selectable). But the textarea IS selectable thus fixing the IE bug.
Also I noticed that it wasn't quite working on Chrome in HTML edit mode due to the transparent color style of addrBubbleHide (the style attached to the hidden element for selecting). I removed it and it seems to work on all browsers without it.

http://bugzilla.zimbra.com/show_bug.cgi?id=56731