Integrate change 279385 from zimbra to zimbra
-------------
Bug 57122
to determine whether the "check all" should be checked we need a specialized treatment in conversastions view, since we only care about
conversations checked, and not messages. To fix it I extracted a _isAllChecked function and override it in ZmConvListView, where it counts conversations selected and compares to the number of conversations.

http://bugzilla.zimbra.com/show_bug.cgi?id=57122