bug 6288,56266,56267
Mailbox Listener rework and clean up.

Added MailboxOperation enum type to denote the current
operation in PendingModification.Change.  This enum
replaces integer based redo operation identifiers,
and also separate redo operation to redo logger
class mapping.

Added pre-modify MailItem map to PendingModification.
The map contains a snapshot of MailItems that were
modified in the current operation.  The snapshot is
taken before any modification was performed, so
the original value is available to the Mailbox
Listeners.

Migrated various calls to RuleManager in Mailbox
code to MailboxListener in order to test the
Listener function.




https://reviewboard.eng.vmware.com/r/215049

http://bugzilla.zimbra.com/show_bug.cgi?id=6288
http://bugzilla.zimbra.com/show_bug.cgi?id=56266
http://bugzilla.zimbra.com/show_bug.cgi?id=56267