bug 6288,56266,56267
Unit test for the MailboxListener.

Change the code to generate the notifications regardless of
any sessions registered for the Mailbox, as we need to
generate notification for static listeners such as
FreeBusyProvider, MemcachedCacheManager, and new FilterListener.

Changed abstract method registerForItemTypes() in the
MailboxListener class to be non abstract method with
default return value for all MailItem types to make
it a little more convenient to create listeners without
creating extra method.

http://bugzilla.zimbra.com/show_bug.cgi?id=6288
http://bugzilla.zimbra.com/show_bug.cgi?id=56266
http://bugzilla.zimbra.com/show_bug.cgi?id=56267