bug: 56642
When an appointment is modified to remove some attendees, set new SEQUENCE to old+2.  Use sequence value of old+1 in the cancel notice sent to removed attendees.  This is consistent with the behavior of MS Outlook and Apple iCal clients.

If the same user receives both update and cancel notices, the greater sequence value in the update will trump the cancel.  If update is received first the later cancel will be ignored.  If cancel is received first the appointment is removed, then re-added when the update is received.

http://bugzilla.zimbra.com/show_bug.cgi?id=56642