bug: 21040
Draft messages are now threaded with their conversation.  If the saveDraft() call comes with an explcitly-specified parent message and that parent is in the draft's mailbox, the draft is slotted into the parent's conversation regardless of the current threading mode.  Otherwise, the mailbox's active threading mode is used to determine a conversation for the draft.

Caveat: Threading is determined when the draft is first saved. Re-saving a draft does *not* update threading.

http://bugzilla.zimbra.com/show_bug.cgi?id=21040