Integrate change 282243 from zimbra to zimbra
-------------
Bug: 58006 - Direct nav to preferences - blank page after save
Rework mgmt of views that get popped when there are no other views underneath, which happens when the user comes via a deep link. Make sure the pop-related callbacks are always called. If the view is not in a tab, then we don't hide it. If it is, and it's compose or msg tab, take the user to mail.

http://bugzilla.zimbra.com/show_bug.cgi?id=58006