Integrate change 282321 from zimbra to zimbra
-------------
Bug: 58052
this is due to part of Bug 56454. It added a margin-left but that is added on top of the 100% width on non-IE, so it becomes more than the available horizontal space.
I replaced it with &nbsp td since that works for all browsers (first I tried
replacing with a padding-left but that worked fine for non IE but did not add the space on IE).
I also noticed that there was no space at all to the left of the invite header in invite messages (organizer, time, invitees, etc), so fixed that with an &nbsp; td as well. Seems fine in all browsers.

http://bugzilla.zimbra.com/show_bug.cgi?id=58052