Integrate change 283719 from HELIX to main
-------------
Bug 57419
if the next item to select is the conversation itself and it's deleted (since
last msg in it is deleted), make sure we don't try to select it since it's
gone. The bug was that we only identified the case the next item to select was within the conv that was deleted, and not the case the next item is the conv itself (since moving backwards after delete of first item, the conv itself is chosen). Also have to include the conv id in the omit array in this case in the recall to _getNextItemToSelect

http://bugzilla.zimbra.com/show_bug.cgi?id=57419