bug: 10545 - support create/edit/import/migrate/sync outlook notes
Added a new system flag called "Note", which can be set on
documents at creation time.  Once a document is created, it
is not possible to change the state of the Note flag.

Updated SOAP and zclient APIs
to support serializing flags for documents.  Added some convenience
methods to MailItem.UnderlyingData for getting/setting the state
of a single flag.  Made UnderlyingData.flags private, to allow us to
validate flag bits each time a change is made.

Added unit tests that verify creating documents with and without
the Note flag set.  Updated HSQLDB database implementation so that
it doesn't attempt to do bitwise operations or LIMIT (MySQL
SQL syntax was incompatible).  Moved server initialization code
for unit tests into a new MailboxTestUtil class.

http://bugzilla.zimbra.com/show_bug.cgi?id=10545