Integrate change 285748 from HELIX to main
-------------
Bug: 55778, 55514 - When reading message, left-click on bubble should select
Created a new class called ZmAddressBubbleList in order to factor out the list and selection handling of bubbles from ZmAddressInputField, so that bubbles in msg headers (handled by the email zimlet) could handle it as well.
The other major part to this was arranging it so that a zimlet could deal with a DwtControl rather than just a SPAN. The control (bubble) handles its events rather than the zimlet framework. That
includes tooltips.
Double-clicking a bubble in a msg header goes to compose, as does single-clicking a highlighted email address in the msg body (or in the header if bubbles are turned off). It used to go to contacts if the address was not in the user's contacts, but now it always goes to compose.

http://bugzilla.zimbra.com/show_bug.cgi?id=55778
http://bugzilla.zimbra.com/show_bug.cgi?id=55514