Integrate change 287834 from HELIX to main
-------------
Bug: 58588 - "On behalf of" not being displayed for messages
Compare actual addresses when trying to figure out if there is a separate sender (which results in "On Behalf Of"). Previously, the template code was comparing zimlet-ified strings after stripping tags (which was a bad idea anyway). That stopped working because the zimlet SPANs for bubbles aren't populated until after the template has been formatted.

http://bugzilla.zimbra.com/show_bug.cgi?id=58588