bug: 58701
Added new soap call SnoozeCalendarItemAlarmRequest.  It works very similarly to the Dismiss- request.  ZWC should call this when user snoozes reminders.

<SnoozeCalendarItemAlarmRequest>
  <appt|task id="cal item id" until="when to show the alarm agian, in millis"/>+
</SnoozeCalendarItemAlarmRequest>

<SnoozeCalendarItemAlarmResponse>
  <appt|task id="cal item id">
    <alarmData ...>  // so the client knows when to trigger the next alarm
  </appt|task>+  // one for each <appt|task> in the request
</SnoozeCalendarItemAlarmResponse>

In <SetAppointmentRequest> and <SetTaskRequest>, the nextAlarm attribute is now understood as possibly snoozed trigger time.  This is compatible with the current behavior of Outlook/ZCO.  No change is needed for clients using SetAppointment/SetTask requests.

http://bugzilla.zimbra.com/show_bug.cgi?id=58701