bug 56269,56270
with this change the children Comments are also deleted when
a Document is deleted.  if dumpster is enabled, the deleted
items are copied to the dumpster.  recovering from dumpster
is still TBD because the current implementation of move to 
dumpster / recover is not symmetric.

made the change to allow the Comments to be put in the dumpster.
mark the children Comments in PendingDelete when Document item is
deleted.  made the change to allow search for MailItem in 
the dumpster table based on parentId.

http://bugzilla.zimbra.com/show_bug.cgi?id=56269
http://bugzilla.zimbra.com/show_bug.cgi?id=56270