bug 56269,56270
changed MailItem.copy() method to take parent MailItem instead of
parentId in order to prevent calling getItemById() to get the
parent MailItem.  getItemById may fail when the parent item is
currently in the dumpster.  with parent MailItem explicitly
passed in the children MailItems can be recovered from 
the dumpster in the same transaction where the parent is being
recovered.

modified Mailbox.copyInternal to recover the children MailItems
in the dumpster in the same transaction.

http://bugzilla.zimbra.com/show_bug.cgi?id=56269
http://bugzilla.zimbra.com/show_bug.cgi?id=56270