Integrate change 297092 from HELIX to main
-------------
Bug: 45618, 58580, 58591 - problems with blank content in HTML editor
Add in a bit of debug logging and some minor possible fixes for problems involving the HTML editor
setting and getting content. The latter is worse, since it can result in the sending of a blank
message without the user knowing it (58580, 58591). The problem with setting content results in a
blank compose area on reply (45618).
One area of suspicion for getting content is _pendingContent, so now we prefer the HTML that is
grabbed from the iframe, unless it's empty or blank (just the HTML wrapper tags).
The problem with setting blank content is possibly a timing issue, so the timeout that is used to
set the content after the iframe has been initialized has been increased from 50ms to 200ms.
- add a "clear" option to reset debug log
- add null check to metadata load handling
- set debug title to HELIX

http://bugzilla.zimbra.com/show_bug.cgi?id=45618
http://bugzilla.zimbra.com/show_bug.cgi?id=58580
http://bugzilla.zimbra.com/show_bug.cgi?id=58591