bug 56269,56270
added support for deleting comments when a folder
containing Document with comments is emptied.

deleted Mailbox.emptySmallFolder() and
Folder.empty() methods.  empty folder will
always use the former implementation of emptyLargeFolder().

simplified the implementation of Folder.delete().
in the previous implementation DeleteScope was not
properly preserved as the scope is changed depending
on whether the folder has subfolders or not.
since Folder.delete() with scope set to CONTENTS_ONLY
is somewhat ambiguous, it's better to force the
caller to use emptyFolder() instead, where the
behavior of keeping the subfolders is specified
explicitly.

http://bugzilla.zimbra.com/show_bug.cgi?id=56269
http://bugzilla.zimbra.com/show_bug.cgi?id=56270