Integrate change 308217 from zimbra to zimbra
-------------
Bug 59904
Fix the errorous codes which use to activate previous/next button.
The misfunction of buttons only occurs in "Member Of" tab, but not in "Members" tab.
As for "Members" tab, after you add the accounts and click save, the buttons
will be enabled if total number of accounts is over 25.
Reviewed:
https://reviewboard.eng.vmware.com/r/231661/	

http://bugzilla.zimbra.com/show_bug.cgi?id=59904