bug 56078 : Add zimbraPrefPop3IncludeSpam to Mail Preferences page
Also fixed a bug with setting as selected the 'Download since now' radio button of the POP_DOWNLOAD_SINCE radio button group.  When there was a 'download since' time specified the setup function mapped it to the incorrect value (a radio button that is hidden).


Review: https://reviewboard.eng.vmware.com/r/231740/

http://bugzilla.zimbra.com/show_bug.cgi?id=56078