Bug 44042
Do not show country flags for languages
Don't show flags for langauge level elements, only locale (country) specific items. Show submenu also for items with just one item in the locale specific item, so that it's consistent (no flag) in the main menu. (e.g. Portuguese (Brazil) )
In addition, show the name of the language/locale in both it's own locale and the user's current locale.

http://bugzilla.zimbra.com/show_bug.cgi?id=44042