bug: 6288 - Event driven third party hook
Fixing the issue of missing pre-modify-items corresponding to some modified and deleted items (particularly folders) in mailbox notifications. We were adding an entry to PendingModifications.preModifyItems map only inside the finally blocks of Mailbox.getItemById() methods. In many scenarios we may not be accessing an item to be modified or deleted using Mailbox.getItemById().

- Removed PendingNotifications.preModifyItems map and instead introduced a preModifyObj field inside PendingModifications.Change.
- Made PendingModifications.deleted map also of type Map<ModificationKey, Change>. For deletions, Change.what would be of type MailItem.Type.
- For deletions, the preModifyObj is populated only when Mailbox.markItemDeleted(MailItem) version of the Mailbox.markItemDeleted() methods is called. This avoids performance overheads in case of bulk delete operations like empty trash.
- Added unit tests.

http://bugzilla.zimbra.com/show_bug.cgi?id=6288