Bug 57258
Ability to remove tag when reading message
add the ability to remove a tag from the list of tags in message header.
I also cleaned up some code that was different for FF. Not sure what were the reasons for it but I tested my common code on FF, Chrome, IE and Safari and it looks ok.

http://bugzilla.zimbra.com/show_bug.cgi?id=57258