Bug 43538 : Extend Snooze Time Durations to Compensate for Longer Reminder Durations
Replaced the Select control with a combination of input field and menu (based on ZmCalItemEditView's reminder drop-down and input combination).  Added the additional snooze periods to the drop down, and added error handling in the SnoozeOKButton handler, since we are now accepting arbitrary input.

Modified the OK button action to send the snooze SOAP request to the server.


Review: https://reviewboard.eng.vmware.com/r/235621/

http://bugzilla.zimbra.com/show_bug.cgi?id=43538