bug 60385
changed the comment activity to be associated to the
document the comment was made to.

changed the activity arguments into a structure in the database
and also in the soap/json response.  this will require
client to change the response parsing to work properly.
currently following arguments are stored for activities.

AddDocumentRevision
  ver - old version
MoveItem
  oldName
  newName
RenameItem
  oldName
  newName
GrantAccess
  target - name of the grantee, or
           "Authuser" for all authenticated users or
                  "Public" for everyone
  role - "Read", "RW", or "Admin"
CreateComment
  parentId
  text

http://bugzilla.zimbra.com/show_bug.cgi?id=60385