Bug 56923
first step in making important right click actions available in the "actions" button in the toolbar. this step is for mark read/unread. Did some refactoring. Also made a change to the actions button to be an
operation like the rest of the buttons. Makes it more consistent and also makes it disabled when user did not choose any item, which is a good thing, since nothing should be enabled in that case as the toolbar is contextual to selected items only.

http://bugzilla.zimbra.com/show_bug.cgi?id=56923