Bug 52483
Provide ability to change color on saved search icons
first part of the implementation of this feature: allow users to choose color when creating a save search. For this I refactored ZmNewSearchDialog to be a subclass of ZmNewOrganizerDialog.  Also required a server-side change to allow for RGB (extended) colors when creating a search folder (support for color was already there, but not the extended rgb colors like in regular folders).

http://bugzilla.zimbra.com/show_bug.cgi?id=52483