Bug 14110 : Add redirect option to mail (sometimes referred to as bounce)
Part 1: because the Redirect UI will need to replicate the 'To' field of the mail message (address input field and bubbles and all), I decided to split the mail message recipient fields (to, cc, bcc) into a separate configurable component (ZmRecipients), to prevent duplication of a lot of code.  This CL handles moving the code out of ZmComposeView into ZmRecipients; a second CL will handle creating the Redirect dialog that uses the ZmRecipients to display the 'To' field.

So ZmRecipients consists mostly of code moved from ZmComposeView, with some changes where callbacks provided by the parent view are used (enableInputs, resizeBody, etc).  The mail.message#Compose template has had the To, CC,and BCC parts removed to separate templates, which are used within the Compose template.



Review : https://reviewboard.eng.vmware.com/r/242491/

http://bugzilla.zimbra.com/show_bug.cgi?id=14110