Bug 61101 : Copy opens dialog with 'undefined' and selected email address not copied
Clipboard copy strings were placed in the wrong file - I moved them.   The bug states that the bubble address was not copied - the dialog (which was missing all text) indicates that the browser does not support automatic copy to clipboard, and provides instructions.


Review : https://reviewboard.eng.vmware.com/r/243338/

http://bugzilla.zimbra.com/show_bug.cgi?id=61101