bug: 61429,16343
Revert to previous code which did not make use of JAXB objects to
construct the results of ZOrganizer.toXML().
When an organizer appears in a SOAP header, it is in the
"urn:zimbra" namespace.  When it appears in the SOAP body, it is
(usually?) in the "urn:zimbraMail" namespace.  The existing JAXB
CalOrganizer object is tied to "urn:zimbraMail".

This issue may need to be revisited as we move more code to use
JAXB - we may choose to change the SOAP definitions to standardise
namespaces for certain encodings in order to avoid multiple, similar
JAXB objects.  See Bug 61406 comment 3 for further discussion of the
issue.

http://bugzilla.zimbra.com/show_bug.cgi?id=61429
http://bugzilla.zimbra.com/show_bug.cgi?id=16343