bug: 6229 - Per-folder message retention policy
Initial checkin for supporting retention policy on folders.
Now storing folder retention policy in the folder's metadata.
The message purge code iterates all the folders in the
mailbox and handles any folders that have retention policy
set.

Added support for getting and setting retention policy via
SOAP:

  <retentionPolicy>?
    <keep>
      <policy type="user" lifetime="{interval}"/>?
      <policy type="system" id="{id}" name="{name}" lifetime="{interval}"/>*
    </keep>
    <purge>
      <policy type="user" lifetime="{interval}"/>?
      <policy type="system" id="{id}" name="{name}" lifetime="{interval}"/>*
    </purge>
  </retentionPolicy>

This block can be passed to <FolderActionRequest> when op="retentionpolicy".
This block is returned under the <folder> element when retention policy
is set, and also in the <modified> block when a folder's retention policy is
modified.

Moved purge unit tests to java-test and added new tests for folder
retention policy.  Added folder retention policy support to ZimbraSoap
and zclient.

To do:

# Limit user policy to 1 element.
# Support tags.
# Admin SOAP API.
# Consolidate server-side and client side RetentionPolicy data structures?
# Put keep/purge policies under a single retentionPolicy parent in Metadata?
# System retention policy.

http://bugzilla.zimbra.com/show_bug.cgi?id=6229