bug: 6229 - Per-folder message retention policy
Consolidated retention policy data structures.  Removed the
server side version.  We now pass the same RetentionPolicy
type to both Mailbox.setRetentionPolicy() and ZFolder.

Moved validation of the message lifetime attribute, to avoid
errors when instantiating folders.  We now check the value
in Mailbox.setRetentionPolicy() and in Mailbox.purgeMessages().
If parsing fails during purge, we log an error and move on.

Disallowed multiple user retention policies per folder.

Updated Mailbox.purgeMessages() to use milliseconds internally
instead of seconds.  This ensures that messages and folders get
deleted by unit tests when lifetime is set to a duration of less than
a second.

http://bugzilla.zimbra.com/show_bug.cgi?id=6229