Bug 60154
IE: After editing the document opens first version
the problem with edit mode is that ZmDocletMgr.prototype.fetchDocumentContent  does not include the
request params (such as ver=version) in the AjxRpc.invoke call. That way it could be getting a cached version (not sure how come it's not happening in other browsers). So I added the ver=version there too. This will kill the cache similar to a random number.

http://bugzilla.zimbra.com/show_bug.cgi?id=60154