bug: 56052 - Default Tags and Folder Tagging
Added support for system-level retention policy.  A user can apply
a system policy to a folder or tag.  The user's policy references
the system policy by id.  If the system policy changes, the user
policy is immediately affected.

System retention policy is stored as metadata in the
zimbraMailPurgeSystemPolicy attribute on globalconfig.
System policy data is cached on the Config object, to
avoid reparsing metadata each time the data is accessed.

Added a <GetSystemRetentionPolicy> mailbox SOAP API, and
also the following admin SOAP APIs for modifying system
retention policy:

<GetSystemRetentionPolicy>
<CreateSystemRetentionPolicy>
<ModifySystemRetentionPolicy>
<DeleteSystemRetentionPolicy>

See soap.txt and soap-admin.txt for the full spec.

http://bugzilla.zimbra.com/show_bug.cgi?id=56052