Bug 59162
Move folder buttons should be like Tag buttons
At first I copy-pasted a lot of the code from ZmChooseFolderDialog and
ZmDialog, but then decided to delegate some using JS "call"... it's a little
awkward, but so is too much duplicate code. This approach also keeps the history and annotations for a lot of code, to help understand what's the purpose of every line.
In case we will remove ZmChooseFolderDialog in the future we can refactor to move the code to DwtFolderChooser.js.
For now the drop-down appears on click on the "move" toolbar button, and on the keyboard shortcut ("M").  on the Actions right-click menu it's still popping up the dialog.

http://bugzilla.zimbra.com/show_bug.cgi?id=59162