Bug 59162
Move folder buttons should be like Tag buttons
disable "move" button on child window. Consistent with previous versions and makes sense to not overload the client on a new window.
this also solves the first issue in bug 61150. (since the current "M" call would exit if the button is not there)

http://bugzilla.zimbra.com/show_bug.cgi?id=59162