Bug 56316
Allow user to select fonts
allow user to select from 4 font options: Modern (the default), Classic, Wide and System. This is saved as ldap pref on the server. On load I attach a class (e.g. user_font_classic) to the body element in order for it to define the font for the entire document. the definition of each of those classes is simply a replacement of property such as @FontFamily-classic@ (this can be overridden in specific skins).

http://bugzilla.zimbra.com/show_bug.cgi?id=56316