Integrate change 324633 from HELIX to main
-------------
Bug 62688
Cannot render message as attachment in a new window when shipped with a calendar invitation
I noticed an inconsistency around ZmInviteMsgView:113. The get is called on the parentAppCtxt if it's from a child window. The account is retrieved from ZmItem.prototype.getAccount (msg.getAccount()) . Inside getAccount, in the "if (this.folderId)", it gets the account from the parentAppCtxt as well. But if that doesn't succeed (or this.folderId is null), it calls ZmOrganizer.parseId(this.id);.  And there, it was getting it always from appCtxt (the child window's appCtxt, instead of parentAppCtxt for a child window).  This was the bug in this case. Hopefully this won't break anything, as I'm not too clear on when the child window should use appCtxt and when parentAppCtxt.

http://bugzilla.zimbra.com/show_bug.cgi?id=62688