Bug 62645
Html tags are not sanitized while changing body format from plain text to html
make sure to convertToHtml when _setBody is called with the baseContent. This would encode HTML characters such as < and &. It was overriding the conversion done by this._htmlEditor.setMode above it, and setting the content as is so if it had html tags, they got into the innerHTML as tags.

http://bugzilla.zimbra.com/show_bug.cgi?id=62645