Bug 63846
Reloading HTTP request URL after Language is changed can include JavaScript and it can be run
also - /res directory in URL is not sanitized
if this problem indeeds happens on IE (I had difficulties trying to repro it on
IE), then it means the <script> "tag" within a JS comment was executed as
script.  That sounds extremely stupid thing to do even for IE. Anyway this
solution removes "<" (they will not legitimately be in the locale name) and
also added a removal of Newlines to avoid the possibility of injecting that and
getting a new script line too.  (at first I tried to use
StringEscapeUtils.escapeHtml  but that didn't seem to make a difference within
a comment).

http://bugzilla.zimbra.com/show_bug.cgi?id=63846