bug:53822 - Support for multiple GAL sync accounts
Take the mailhost/server as mandatory parameter in CreateGalSyncAccount and restrict one zimbra galsync account per server per domain.
Change the CLI to take this new argument.
Add new SOAP API "AddGalSyncDataSourceRequest" for adding additional datasources to the existing galsync account.
For autocomplete requests, give the preference to the galsync account present on the local server if one exists.
For SyncGal requests, check if the client was already synced with a galsync account. Use the same galsync account for subesequent SyncGal requests. 

http://bugzilla.zimbra.com/show_bug.cgi?id=53822