bug: 58159
Added special case to zmrcd script to handle server-mode rsync invocation over ssh from a remote rsync initiator host.  The SSH_ORIGINAL_COMMAND environment variable is examined to see if the command being invoked is rsync.  Only "rsync" and "/opt/zimbra/rsync/bin/rsync" are allowed.  Furthermore, the command is allowed only when the entire string satisfies the args regex, which is defined as '[\'"a-zA-Z0-9\\/\\.\\-_:@ ]*'.  When all conditions are met zmrcd exec's the original command.

http://bugzilla.zimbra.com/show_bug.cgi?id=58159