Bug 64833
Cannot open message using a deep-link to a message (view=msg&id=xxx)
the session ID can be a Number (in this case, 1. see
ZmApp.prototype.getSessionController), so indexOf would not work. Making sure indexOf would fail (by calling toString(), so in case of a Number the string we look for would not be found) fixes the problem at hand.

http://bugzilla.zimbra.com/show_bug.cgi?id=64833