bug 64579 item level sharing / grant / revoke support.
moved the grant / revoke functionality from Folder to MailItem.
grant now becomes intrinsic member for all the mail items.
grant is initialized only for Document and Folders for now.
as we add additional support for different items we can change
it as needed.

this checkin does not include the publishing of iteml level
grants to LDAP.  filed bug 65057 for this enhancement.

currently sharing document supports sharing to single user
or public only.  this could be changed to support additional 
principal types later on.

grant on the Document can be added by

    <DocumentActionRequest xmlns="urn:zimbraMail">
      <action id="260" op="grant">
        <grant gt="pub" perm="r"/>
      </action>
    </DocumentActionRequest>

then the item can be accessed by the URL

    http://<server>/service/home/<username>?id=260

https://reviewboard.eng.vmware.com/r/267224

http://bugzilla.zimbra.com/show_bug.cgi?id=64579