Bug 42915
End users should be able to view lists they already belong to and unsubscribe
approval workflow - special approval request (for subscribe/unsubscribe) is sent to the owners. If message has dlSubs component, show the approve/reject buttons toolbar. Send appropriate request to the server when this owner approves or rejects.

http://bugzilla.zimbra.com/show_bug.cgi?id=42915