bug: 63526
Fix bad item_id_checkpoint during migration. This script is intended as a stop-gap to help isolate the root cause of bug 46549.

The script puts mailbox back into a state where ZD and ZCS no longer use the same ID range for new items. It does not remove existing items which were previously created in the bad range, so item_id_checkpoint is not necessarily the highwater mark for mail items when mailbox has been corrupted with bad IDs. It is typically necessary to reset ZD account after the item_id_checkpoint has been corrected.

If a mailbox has an invalid item_id_checkpoint and a gap of at least 10 million items cannot be found the item_id_checkpoint will not be altered. This case would represent a severly corrupted mailbox, and a complete export/delete/recreate/import would be necessary. 

A complete rebuild may also be necessary if there are a large number of bad IDs in the ZCS mailbox. 

http://bugzilla.zimbra.com/show_bug.cgi?id=63526