Bug 64894
Not Spam button doesn't work properly when searching Junk folder
no need to calculate twice if this is a spam or "not spam" action. We already
calculated it (correctly) when updating the button text, so keep that
information in easy way (flag) with the button to be used by the listener.

http://bugzilla.zimbra.com/show_bug.cgi?id=64894