Bug 64857 : Getting repeated calendar reminders for same event
After discussion with Josh and Greg, here is the desired behavior:
- The 'before' snooze periods will be limited by the time period to the next future appt's start time (i.e. appts that are in-progress or completed are disregarded).  The dialog will contain the next smallest standard entry (so if its 8 minutes before a meeting, the first entry will be 5 minutes before).
- The fixed snooze periods will be limited by the longest time to the start of any of the appts in the reminder list. 
- The default will be the longest 'before' entry.

The snooze menu will be repopulated via the housekeeping loop, so that menu choices change if the reminder dialog is left open.

When applying the snooze period:
- If fixed, a new reminder time is calculated from now and applied to all the appts.
- If before, all future appts have the before period applied to their start times individually.  The soonest of these will also be applied to any appts that are in-progress or completed.  The in-progress and completed appts should be dismissed, but if not this prevents them from immediately popping up the reminder dialog again.

If the reminder dialog contains only appts whose start dates are in the past (in-progress or completed appts) then the snooze input field, menu and button are disabled.

Also fixed the formatting fencepost problem - i.e. the snooze menu was showing '60 minutes' instead of '1 hour', '24 hours' instead of '1 day' etc.


Review : https://reviewboard.eng.vmware.com/r/268706/

http://bugzilla.zimbra.com/show_bug.cgi?id=64857