bug: 16343
Create a single WSDL file instead of one for each namespace.
Could potentially decide to have 2 WSDL files, one for Admin and
one for everything else.  If we do, that will be easier after
these changes.
Has the useful side effect of halving the number of generated
files needed for wsdl-test. Plan to submit changes to
generated files separately.

https://reviewboard.eng.vmware.com/r/273723/

http://bugzilla.zimbra.com/show_bug.cgi?id=16343