Bug 66068
Empty Distribution Lists: tooltip shows "Contacts: 1"
For now I think it's fine to not have the tooltip for the DL folder. This does that.
I also fixed a bug I noticed that was causing the "actions" toolbar button to not be enabled when moving from any GAL entry (including DL) to a regular contact. This was due to the ZmContactListController.prototype._resetOperations
not calling ZmBaseController.prototype._resetOperations so once the "actions" button was disabled, it was never re-enabled.

http://bugzilla.zimbra.com/show_bug.cgi?id=66068