bug: 65377 - Handle and localize known server errors
Added the SoapFaultHandler framework, which is used
for handling the most common SOAP faults.  When the
server returns a SOAP fault, we compare the code and
arguments against the list of known faults.  If there's a
match, we display the localized, friendly error message
to the user.  If there's no match, we display the unlocalized
error that comes from the server.

Cleaned up SoapFaultException, so that it passes the
arguments to the superclass constructor.  This allows
SoapFaultException.getArgs() to return the same data
as the original server-side ServiceException.

Updated ServiceException.getArgs() to return a
List instead of an array, for more flexibility.  Updated the
ServiceException internals so that the args list and
the values of each Argument are not modifiable.

Updated the method signature of
MailServiceException.ALREADY_EXISTS so that the
arguments are clearly defined.

http://bugzilla.zimbra.com/show_bug.cgi?id=65377