Bug 65297
both browser context menu and zimbra context menu show for dates
The context menu is shown in the oncontextmenu
event of the new FF.  Not mouseDown like it was in older versions (and possibly still is on Mac). Also not on mouseUp like it may look when holding the button and releasing it.
To solve it I did the same thing we do for Safari, which is a contextmenu
listener that delegates to ZmObjectManager.prototype._mouseDownListener. This
turns out to work.
I also need to ignore the capture on right-click in
DwtIframe.prototype._rawEventHandler so it notifies the listeners.

http://bugzilla.zimbra.com/show_bug.cgi?id=65297