bug: 38813
Fix sending on behalf of pop/imap/caldav data sources.  Allow send-obo if the From address is the address of a pop/imap/caldav data source, the data source is enabled, and the address is an external address.  Internal addresses are excluded because sending on behalf of an internal address requires the sendOnBehalfOf right to be granted.

http://bugzilla.zimbra.com/show_bug.cgi?id=38813