bug: 66629, 67191
Use redirect mode when sending auto-forwarded invite email from delegator to delegate.  The forwarded email has X-Zimbra-Calendar-Intended-For header indicating who is doing the forward.  Without redirect mode the Sender header is also set to the forwarder, turning this into a send-on-behalf-of.  The message is still valid, but when the UI displays this message as send-obo it can be confusing to the user.

http://bugzilla.zimbra.com/show_bug.cgi?id=66629
http://bugzilla.zimbra.com/show_bug.cgi?id=67191