bug 67360 - uploading files to shared folder shows incorrect activity in history
added class member PendingModifications.Change.who to capture
the account information performing the transaction.  this needs
to be captured in the audit log and activity stream.

also changed the Map PendingModifications.created to store Change
object as the value instead of MailItem object, so that we
keep track of the account that created the object.

https://reviewboard.eng.vmware.com/r/282431

http://bugzilla.zimbra.com/show_bug.cgi?id=67360