Integrated the following from //depot/zimbra/HELIX/ZimbraServer/conf/jetty-6.1/...: Change 341190 by dcampbell@dcampbell_dcampbell-MBPRO on 2011/12/02 12:19:54
Bug: 67171

Add attribute to Jetty.xml to block SSL renegotiate.  Jetty documentation shows that this is the default behavior, yet this is in fact inaccurate.  Default behavior as of Jetty changeid 6757 is to allow SSL renegotiations and let JSSE handle whether to allow renegotiation or not.  JSSE can only block "unsafe" renegotatons, while with SslSelectChannelConnector we can be much more specific and block ALL renegotations, to avoid DOS renegotiation attacks.

http://bugzilla.zimbra.com/show_bug.cgi?id=67171