bug: 22901
Moved MTA connection check for cancel notice out of the mail send layer.  The check is now done earlier in the request handler.  If no MTA is available, don't allow the organizer to cancel an appointment with attendees.  This prevents orphaning the attendees.

Introduced a mail queue for sending calendar notice emails.  Queue the emails instead of sending them inline during appointment update.  Send the queued emails after the updates are finished and mailbox lock is released.  The send happens in the request handlers.

http://bugzilla.zimbra.com/show_bug.cgi?id=22901